# Rust compile time optimizations