--- a/app/gui/language/ast/macros/src/lib.rs
+++ b/app/gui/language/ast/macros/src/lib.rs
@@ -14,12 +14,11 @@ extern crate proc_macro;
 
 mod token;
 
-use crate::prelude::*;
 use crate::token::TokenDescription;
 
+use std::collections::HashSet;
 use enso_macro_utils::gather_all_type_reprs;
 use enso_macro_utils::repr;
-use enso_prelude as prelude;
 use proc_macro2::Ident;
 use proc_macro2::Span;
 use proc_macro2::TokenStream;
@@ -110,7 +109,7 @@ fn mk_product_type(
     let ident_nested = format!("{}{}", decl.ident, variant.ident);
     let ident_nested = Ident::new(&ident_nested, Span::call_site());
     let ident = if is_flat { ident_flat } else { ident_nested };
-    let generics = syn::Generics { params, ..default() };
+    let generics = syn::Generics { params, ..Default::default() };
     let mut fields = variant.fields.clone();
     let semi_token = None;
     fields.iter_mut().for_each(|f| f.vis = vis.clone());