--- a/app/gui/language/ast/macros/src/lib.rs
+++ b/app/gui/language/ast/macros/src/lib.rs
@@ -16,13 +16,13 @@ mod token;
 
 use crate::token::TokenDescription;
 
-use std::collections::HashSet;
 use enso_macro_utils::gather_all_type_reprs;
 use enso_macro_utils::repr;
 use proc_macro2::Ident;
 use proc_macro2::Span;
 use proc_macro2::TokenStream;
 use quote::quote;
+use std::collections::HashSet;