--- a/lib/rust/ensogl/component/text/src/font/family/src/lib.rs
+++ b/lib/rust/ensogl/component/text/src/font/family/src/lib.rs
@@ -49,6 +49,7 @@ use derive_more::Display;
 use std::collections::HashMap;
 use std::fmt::Formatter;
 
+
 // ==============
 // === Export ===
 // ==============