--- a/lib/rust/ensogl/component/text/src/font/family/src/lib.rs
+++ b/lib/rust/ensogl/component/text/src/font/family/src/lib.rs
@@ -47,7 +47,6 @@
 use derive_more::Deref;
 use derive_more::Display;
 use std::collections::HashMap;
-use std::fmt::Formatter;
 
 
 // ==============