--- a/lib/rust/frp/src/network.rs
+++ b/lib/rust/frp/src/network.rs
@@ -62,7 +62,7 @@ pub struct NetworkData {
 
 impl NetworkData {
     /// Constructor.
-    pub fn new(label: impl Into<String>) -> Self {
+    pub fn new(label: String) -> Self {
         let label = label.into();
         let nodes = default();
         let links = default();
@@ -81,6 +81,11 @@ impl Drop for NetworkData {
 impl Network {
     /// Constructor.
     pub fn new(label: impl Into<String>) -> Self {
+        Self::new_with_string(label.into())
+    }
+
+    /// Non-generic constructor.
+    fn new_with_string(label: String) -> Self {
         let data = Rc::new(NetworkData::new(label));
         Self { data }
     }
@@ -99,23 +104,33 @@ impl Network {
     /// network, like animation instances.
     pub fn store<T: 'static + CloneRef>(&self, item: &T) {
         let item = item.clone_ref();
-        self.data.storage.borrow_mut().push(Box::new(item));
+        self.store_boxed(Box::new(item));
+    }
+
+    /// Store an arbitrary boxed item.
+    #[inline(never)]
+    fn store_boxed(&self, item: Box<dyn Any>) {
+        self.data.storage.borrow_mut().push(item);
     }
 
     /// Register the node and return it's weak reference.
     pub fn register_raw<T: HasOutputStatic>(&self, node: stream::Node<T>) -> stream::WeakNode<T> {
         let weak = node.downgrade();
-        let node = Box::new(node);
-        self.data.nodes.borrow_mut().push(node);
+        self.register_boxed(Box::new(node));
         weak
     }
 
     /// Register the node and return a new `Stream` reference.
     pub fn register<Def: HasOutputStatic>(&self, node: stream::Node<Def>) -> Stream<Output<Def>> {
         let stream = node.clone_ref().into();
-        let node = Box::new(node);
-        self.data.nodes.borrow_mut().push(node);
+        self.register_boxed(Box::new(node));
         stream
+    }
+
+    /// Register a boxed node. Static part of registration.
+    #[inline(never)]
+    fn register_boxed(&self, node: Box<dyn Item>) {
+        self.data.nodes.borrow_mut().push(node);
     }
 
     /// Register a new link between nodes. Visualization purposes only.