### Pull Request Description

Rust compile time optimizations, tracked by https://www.pivotaltracker.com/story/show/183746783

### Important Notes

The majority of time compile optimization in this PR comes from changes in dependency tree. Some dependency duplication got removed, some heavy to compile features got disabled (especially from zip crate). The `prelude` crate is also now compiled only for wasm target. Its (very minimal) usage from our proc-macro crates got removed.

This means the biggest changes will be visible in full rebuild, and significant part of it will also reduce partial builds after modifying anything in prelude or its dependencies.


| | before |  after |
|-|-|-|
|total build time|3m 49.7s|3m 27.6s|
| total compilation units | 909 | 766 |
|ide-view-graph-editor build time| 44.9s | 43.5s |
|ide-view-component-list-panel-grid build time| 30.3s | 28.0s |
|enso_gui.wasm file size (no wasm-opt)| 80.9 MB | 80.3 MB |
|enso_gui.wasm file size (wasm-opt -O3)| 65.4 MB | 65.0 MB |

The frp callstack handling changes should also have positive runtime performance implications when callstacks are enabled. But it's hard to measure that in isolation. It's at least clear that the callstack "sub" method basically disappeared from profiles. Very rougly comparing the opening of component browser (whole event takes ~1s) - before "sub" took 42.4ms, now it takes 0.4ms.


This is only part of the work from the compile time optimization effort - extracted the bits that are most impactful and with safest code changes. A lot is still on the table in the FRP framework core, but for that further work will be needed. So far, the changes there don't make enough difference to consider them for merging.

### Checklist

Please include the following checklist in your PR:

- [x] The documentation has been updated if necessary.
- [x] All code conforms to the
      [Scala](https://github.com/enso-org/enso/blob/develop/docs/style-guide/scala.md),
      [Java](https://github.com/enso-org/enso/blob/develop/docs/style-guide/java.md),
      and
      [Rust](https://github.com/enso-org/enso/blob/develop/docs/style-guide/rust.md)
      style guides.
- All code has been tested:
  - [x] Unit tests have been written where possible.
  - [x] If GUI codebase was changed: Enso GUI was tested when built using BOTH
        `./run ide build` and `./run ide watch`.